Use a default export without default keyword
