+++ package.json
@@ -3,10 +3,10 @@
   "version": "1.23.0",
   "description": "Common repository for example datasets used by Vega related projects.",
   "main": "build/vega-datasets",
-  "module": "build/index",
+  "module": "build/data",
   "unpkg": "build/vega-datasets.min.js",
   "jsdelivr": "build/vega-datasets.min.js",
-  "typings": "build/index.d.ts",
+  "typings": "build/data.d.ts",
   "repository": {
     "type": "git",
     "url": "http://github.com/vega/vega-datasets.git"
@@ -16,13 +16,13 @@
     "url": "http://idl.cs.washington.edu"
   },
   "devDependencies": {
-    "@types/d3-dsv": "^1.0.35",
+    "@types/d3-dsv": "^1.0.36",
     "d3-dsv": "^1.1.1",
     "datalib": "^1.9.2",
-    "rollup": "^1.6.0",
+    "rollup": "^1.7.3",
     "rollup-plugin-node-resolve": "^4.0.1",
     "terser": "^3.17.0",
-    "typescript": "^3.3.3"
+    "typescript": "^3.3.4000"
   },
   "scripts": {
     "index": "./scripts/make-url-index.sh > src/urls.ts",
